+++ src/repo/mongoose/model/task.ts
@@ -2,24 +2,6 @@ import * as mongoose from 'mongoose';
 
 const writeConcern: mongoose.WriteConcern = { j: true, w: 'majority', wtimeout: 10000 };
 
-const executionResultSchema = new mongoose.Schema(
-    {},
-    {
-        id: false,
-        _id: false,
-        strict: false
-    }
-);
-
-const dataSchema = new mongoose.Schema(
-    {},
-    {
-        id: false,
-        _id: false,
-        strict: false
-    }
-);
-
 /**
  * タスクスキーマ
  */
@@ -32,8 +14,8 @@ const schema = new mongoose.Schema(
         remainingNumberOfTries: Number,
         lastTriedAt: Date,
         numberOfTried: Number,
-        executionResults: [executionResultSchema],
-        data: dataSchema
+        executionResults: [mongoose.SchemaTypes.Mixed],
+        data: mongoose.SchemaTypes.Mixed
     },
     {
         collection: 'tasks',
@@ -46,8 +28,18 @@ const schema = new mongoose.Schema(
             createdAt: 'createdAt',
             updatedAt: 'updatedAt'
         },
-        toJSON: { getters: true },
-        toObject: { getters: true }
+        toJSON: {
+            getters: false,
+            virtuals: false,
+            minimize: false,
+            versionKey: false
+        },
+        toObject: {
+            getters: false,
+            virtuals: true,
+            minimize: false,
+            versionKey: false
+        }
     }
 );