+++ src/repo/mongoose/schema/offer.ts
@@ -51,8 +51,18 @@ export function create(options: mongoose.SchemaOptions) {
                 createdAt: 'createdAt',
                 updatedAt: 'updatedAt'
             },
-            toJSON: { getters: true },
-            toObject: { getters: true }
+            toJSON: {
+                getters: false,
+                virtuals: false,
+                minimize: false,
+                versionKey: false
+            },
+            toObject: {
+                getters: false,
+                virtuals: true,
+                minimize: false,
+                versionKey: false
+            }
         }
     );